fix: sidebar always match "/"
